r Tidy packages exports
