+++ packages/core/src/compiler/addons/index.ts
@@ -4,9 +4,6 @@
  *   Licensed under the MIT License. See LICENSE in the project root for license information.
  * ---------------------------------------------------------------------------------------------
  */
-import { createResolver } from "./addon-resolver";
-import { AddonRegistry, type AddonConfig } from "./AddonRegistry";
-import { compilerAddons, type CompilerAddon, type CompilerAddons } from "./CompilerAddon";
-
-export { AddonRegistry, compilerAddons, createResolver };
-export type { AddonConfig, CompilerAddon, CompilerAddons };
+export { createResolver } from "./addon-resolver";
+export { AddonRegistry } from "./AddonRegistry";
+export type { CompilerAddon, CompilerAddons } from "./CompilerAddon";
