+++ packages/core/src/compiler/cache/index.ts
@@ -4,8 +4,5 @@
  *   Licensed under the MIT License. See LICENSE in the project root for license information.
  * ---------------------------------------------------------------------------------------------
  */
-import { CacheFile, getCachedName } from "./CacheFile";
-import { FileCache } from "./FileCache";
-
-export type { CacheFile };
-export { getCachedName, FileCache };
+export { getCachedName, type CacheFile } from "./CacheFile";
+export { FileCache } from "./FileCache";
