+++ packages/core/src/compiler/compilation/index.ts
@@ -4,8 +4,6 @@
  *   Licensed under the MIT License. See LICENSE in the project root for license information.
  * ---------------------------------------------------------------------------------------------
  */
-import { CompilationContext } from "./CompilationContext";
-import { CompilationHost } from "./CompilationHost";
-import { createSharedHost } from "./shared-host";
-
-export { createSharedHost, CompilationContext, CompilationHost };
+export { CompilationContext, type CompilationContextOptions } from "./CompilationContext";
+export { CompilationHost } from "./CompilationHost";
+export { createSharedHost } from "./shared-host";
