+++ packages/core/src/compiler/config/index.ts
@@ -4,10 +4,7 @@
  *   Licensed under the MIT License. See LICENSE in the project root for license information.
  * ---------------------------------------------------------------------------------------------
  */
-import type { CompilationConfig } from "./CompilationConfig";
-import { resolveCompilationConfig, updateCompilerOptions } from "./resolve-compiler-config";
-import { resolveProjectConfig } from "./resolve-project-config";
-import { resolveTargets } from "./resolve-targets";
-
-export { resolveProjectConfig, resolveCompilationConfig, resolveTargets, updateCompilerOptions };
-export type { CompilationConfig };
+export type { CompilationConfig } from "./CompilationConfig";
+export { resolveCompilationConfig, updateCompilerOptions } from "./resolve-compiler-config";
+export { resolveProjectConfig } from "./resolve-project-config";
+export { resolveTargets } from "./resolve-targets";
