+++ packages/webpack/src/index.ts
@@ -5,33 +5,7 @@
  * ---------------------------------------------------------------------------------------------
  */
 
-import { LoaderContext } from "webpack";
-import { initializeInstance, setInstanceInCache } from "./instance-cache";
-import { getLoaderOptions, PluginOptions } from "./loader-options";
-import { processResultAndFinish } from "./result-handling";
-import { TsCompiler } from "./TsCompiler";
-
-function loader(this: LoaderContext<PluginOptions>): void {
-    this.cacheable && this.cacheable();
-    const loaderOptions = getLoaderOptions(this);
-    const instance = initializeInstance(this, loaderOptions, (path: string) => {
-        this.addDependency(path);
-    });
-    const fragment = buildTargets(instance, this.resourcePath);
-
-    this.version = instance.version;
-
-    setInstanceInCache(this._compiler, this, instance);
-
-    processResultAndFinish(this, fragment, instance.targets);
-}
-
-const buildTargets = (compiler: TsCompiler, resourcePath: string) => {
-    return compiler.build(resourcePath);
-};
-
-export default loader;
-
-export { Compiler, createBrowserSystem, DefaultReporter, getVersionedFile, NoReporter } from "@quatico/websmith-core";
-export { Upath as uPath } from "./Upath";
+export { Compiler, DefaultReporter, NoReporter, createBrowserSystem, getVersionedFile } from "@quatico/websmith-core";
 export type { CompilerOptions } from "@quatico/websmith-core";
+export { Upath as uPath } from "./Upath";
+export { loader as default } from "./loader";
